准备自己跑rq1a b，但是遇到memo err的最新版本

+++ load_model/get_pre_dataset.py
@@ -1,3 +1,4 @@
+import os.path
 import sys
 
 sys.path.append("../")
@@ -9,8 +10,8 @@ from tutorial_models import *
 dataset_rela_list = ['adult', 'bank', 'compas', 'default',
                      'german', 'heart', 'meps15', 'meps16',
                      'student']
-dataset_abso_list = ['adult', 'compas', 'bank',
-                     'default', 'meps15', 'meps16']  # german heart student 基数太小，没有绝对的数据集可以选择
+dataset_abso_list = ['adult',
+                     'default', 'meps15', 'meps16']  # 变成4个 5组
 dataset_shape_map = {
     'adult': (None, 14), 'bank': (None, 20),
     'compas': (None, 21), 'german': (None, 20),
@@ -18,9 +19,11 @@ dataset_shape_map = {
     'student': (None, 32),
     'meps15': (None, 42), 'meps16': (None, 42)}
 
+model_type_list= ['dnn1', 'dnn2', 'dnn3', 'dnn4', 'dnn5']
 
 def predict_dnn5(sample_feed_path, input_shape=(None, 20), nb_classes=2,
-                 model_path='./bank-additional/model/bank-additional/999/' + 'test.model'):
+                 model_path='./bank-additional/model/bank-additional/999/' + 'test.model',
+                 model_type='dnn5'):
     """
     :param sample_feed_path: 输入的测试集npy文件所在路径
     :param input_shape: 由模型指定
@@ -36,8 +39,28 @@ def predict_dnn5(sample_feed_path, input_shape=(None, 20), nb_classes=2,
     sess = tf.Session(config=config)
     x = tf.placeholder(tf.float32, shape=input_shape)
     y = tf.placeholder(tf.float32, shape=(None, nb_classes))
-    model = dnn5(input_shape, nb_classes)
-    preds = model(x)
+    if model_type == 'dnn1':
+        model = dnn1(input_shape, nb_classes)
+        preds = model(x)
+    elif model_type == 'dnn3':
+        model = dnn3(input_shape, nb_classes)
+        preds = model(x)
+    elif model_type == 'dnn7':
+        model = dnn7(input_shape, nb_classes)
+        preds = model(x)
+    elif model_type == 'dnn9':
+        model = dnn9(input_shape, nb_classes)
+        preds = model(x)
+    elif model_type == 'dnn2':
+        model = dnn2(input_shape, nb_classes)
+        preds = model(x)
+    elif model_type == 'dnn4':
+        model = dnn4(input_shape, nb_classes)
+        preds = model(x)
+    else:
+        assert model_type == 'dnn5'
+        model = dnn5(input_shape, nb_classes)
+        preds = model(x)
     grad_0 = gradient_graph(x, preds)
     # saver = tf.train.Saver()
     model_path = model_path
@@ -58,7 +81,7 @@ def predict_dnn5(sample_feed_path, input_shape=(None, 20), nb_classes=2,
     return label_tmp, ranker_score
 
 
-def gen_relative_sample_feed_path_list(dataset_name='adult', test_id=1):
+def gen_relative_sample_feed_path_list(dataset_name='adult', test_id=1, model_type='dnn5'):
     two_d_labels_test_rela_path_list = []
     features_test_rela_path_list = []
     labels_test_rela_path_list = []
@@ -71,15 +94,19 @@ def gen_relative_sample_feed_path_list(dataset_name='adult', test_id=1):
         feat_path = '../data/' + dataset_name + '-aif360preproc-done/' + 'test/' + str(
             test_id) + '/featrures-test-' + str((i + 1) * 5) + '%.npy'
         pred_path = '../data/' + dataset_name + '-aif360preproc-done/' + 'test/' + str(
-            test_id) + '/2d-score-test-' + str((i + 1) * 5) + '%.npy'
+            test_id) + '/' + model_type + '/' + '2d-score-test-' + str((i + 1) * 5) + '%.npy'
         two_d_labels_test_rela_path_list.append(two_d_path)
         labels_test_rela_path_list.append(label_path)
         features_test_rela_path_list.append(feat_path)
         pred_rela_path_list.append(pred_path)
+    if not os.path.exists('../data/' + dataset_name + '-aif360preproc-done/' + 'test/' + str(
+            test_id) + '/' + model_type + '/'):
+        os.makedirs('../data/' + dataset_name + '-aif360preproc-done/' + 'test/' + str(
+            test_id) + '/' + model_type + '/')
     return two_d_labels_test_rela_path_list, features_test_rela_path_list, labels_test_rela_path_list, pred_rela_path_list
 
 
-def gen_absolute_sample_feed_path_list(dataset_name='adult', test_id=1):
+def gen_absolute_sample_feed_path_list(dataset_name='adult', test_id=1, model_type='dnn5'):
     two_d_labels_test_abso_path_list = []
     features_test_abso_path_list = []
     labels_test_abso_path_list = []
@@ -92,11 +119,15 @@ def gen_absolute_sample_feed_path_list(dataset_name='adult', test_id=1):
         feat_path = '../data/' + dataset_name + '-aif360preproc-done/' + 'test/' + str(
             test_id) + '/featrures-test-' + str((i + 1) * 100) + '.npy'
         pred_path = '../data/' + dataset_name + '-aif360preproc-done/' + 'test/' + str(
-            test_id) + '/2d-score-test-' + str((i + 1) * 100) + '.npy'
+            test_id) + '/' + model_type + '/' + '2d-score-test-' + str((i + 1) * 100) + '.npy'
         two_d_labels_test_abso_path_list.append(two_d_path)
         labels_test_abso_path_list.append(label_path)
         features_test_abso_path_list.append(feat_path)
         pred_abso_path_list.append(pred_path)
+    if not os.path.exists('../data/' + dataset_name + '-aif360preproc-done/' + 'test/' + str(
+            test_id) + '/' + model_type + '/'):
+        os.makedirs('../data/' + dataset_name + '-aif360preproc-done/' + 'test/' + str(
+            test_id) + '/' + model_type + '/')
     return two_d_labels_test_abso_path_list, features_test_abso_path_list, labels_test_abso_path_list, pred_abso_path_list
 
 
@@ -113,37 +144,48 @@ def all_model_path_list(model_type='dnn5'):
     return path_list
 
 
-
 def get_rela_pred_npy():
-    for dataset_id in range(len(dataset_rela_list)):
-        print('current dataset is %s' % dataset_rela_list[dataset_id])
-        for test_id in range(5):
-            print('current test id is %d' % (test_id +1))
-            two_d_labels_test_rela_path_list, features_test_rela_path_list, labels_test_rela_path_list, pred_rela_path_list= \
-                gen_relative_sample_feed_path_list(dataset_rela_list[dataset_id], test_id + 1)
-            for test_path_i in range(len(features_test_rela_path_list)):
-                model_path = gen_model_path(dataset_name=dataset_rela_list[dataset_id], model_type='dnn5')
-                label_tmp, ranker_score = predict_dnn5(features_test_rela_path_list[test_path_i],
-                             input_shape=dataset_shape_map[dataset_rela_list[dataset_id]],
-                             nb_classes=2,
-                             model_path=model_path)
-                np.save(pred_rela_path_list[test_path_i], ranker_score)
+    for j in range(len(model_type_list)):
+        model_type = model_type_list[j]
+        for dataset_id in range(len(dataset_rela_list)):
+            print('>>>>get rela<<<<')
+            print('current dataset is %s' % dataset_rela_list[dataset_id])
+            for test_id in range(5):
+                print('current test id is %d' % (test_id + 1))
+                two_d_labels_test_rela_path_list, features_test_rela_path_list, labels_test_rela_path_list, pred_rela_path_list = \
+                    gen_relative_sample_feed_path_list(dataset_rela_list[dataset_id], test_id + 1, model_type=model_type)
+                for test_path_i in range(len(features_test_rela_path_list)):
+                    model_path = gen_model_path(dataset_name=dataset_rela_list[dataset_id], model_type=model_type)
+                    if not os.path.exists(pred_rela_path_list[test_path_i]):
+                        label_tmp, ranker_score = predict_dnn5(features_test_rela_path_list[test_path_i],
+                                                               input_shape=dataset_shape_map[
+                                                                   dataset_rela_list[dataset_id]],
+                                                               nb_classes=2,
+                                                               model_path=model_path,
+                                                               model_type=model_type)
+                        np.save(pred_rela_path_list[test_path_i], ranker_score)
+
 
 
 def get_abso_pre_npy():
-    for i in range(len(dataset_abso_list)):
-        print('current dataset is %s' % dataset_abso_list[i])
-        for j in range(5):
-            print('current test id is %d' % (j+1))
-            two_d_labels_test_abso_path_list, features_test_abso_path_list, labels_test_abso_path_list, pred_abso_path_list = \
-                gen_absolute_sample_feed_path_list(dataset_abso_list[i], j+1)
-            for path_i in range(len(features_test_abso_path_list)):
-                model_path = gen_model_path(dataset_name=dataset_abso_list[i], model_type='dnn5')
-                label_tmp, ranker_score = predict_dnn5(features_test_abso_path_list[path_i],
-                                           input_shape=dataset_shape_map[dataset_abso_list[i]],
-                                           nb_classes=2,
-                                           model_path=model_path)
-                np.save(pred_abso_path_list[path_i], ranker_score)
+    for k in range(len(model_type_list)):
+        model_type = model_type_list[k]
+        for i in range(len(dataset_abso_list)):
+            print('====get abso ===')
+            print('current dataset is %s' % dataset_abso_list[i])
+            for j in range(5):
+                print('current test id is %d' % (j + 1))
+                two_d_labels_test_abso_path_list, features_test_abso_path_list, labels_test_abso_path_list, pred_abso_path_list = \
+                    gen_absolute_sample_feed_path_list(dataset_abso_list[i], j + 1, model_type=model_type)
+                for path_i in range(len(features_test_abso_path_list)):
+                    model_path = gen_model_path(dataset_name=dataset_abso_list[i], model_type=model_type)
+                    if not os.path.exists(pred_abso_path_list[path_i]):
+                        label_tmp, ranker_score = predict_dnn5(features_test_abso_path_list[path_i],
+                                                           input_shape=dataset_shape_map[dataset_abso_list[i]],
+                                                           nb_classes=2,
+                                                           model_path=model_path,
+                                                           model_type=model_type)
+                        np.save(pred_abso_path_list[path_i], ranker_score)
 
 
 if __name__ == '__main__':
